requirements.txt
markdown2
matplotlib
Pillow
pyperclip
requests
tkinterweb
